CortexAI/transformationAPI/component/url/urls.py
from django.urls import path
from django.views.generic import TemplateView
from transformationAPI.component.view import *
from rest_framework import permissions
from drf_yasg.views import get_schema_view
from drf_yasg import openapi

schema_view = get_schema_view(
   openapi.Info(
      title="CoretexAI API",
      default_version='v1',
      description="Test description",
      terms_of_service="https://www.google.com/policies/terms/",
      contact=openapi.Contact(email="pacloutieru@deloitte.ca"),
      license=openapi.License(name="BSD License"),
   ),
   public=True,
   permission_classes=(permissions.AllowAny,),
)

urlpatterns = [
    path('aws_athena_databases', aws_athena_databases.Databases.as_view(),
         name='aws_athena_databases'),
    path('swagger',
         schema_view.with_ui('swagger', cache_timeout=0),
         name='schema-swagger-ui'),
]
