requirements.txt
asgiref==3.4.1
asn1crypto==1.4.0
awswrangler==2.11.0
beautifulsoup4==4.10.0
boto3==1.18.40
botocore==1.21.40
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.2
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
pg8000==1.21.1
psycopg2==2.9.1
pyarrow==5.0.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
redshift-connector==2.0.886
requests==2.25.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scramp==1.4.1
simplejson==3.17.5
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6